prescription_reader/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow>=9.0.0
pdf2image==1.17.0
groq==0.9.0
httpx==0.27.0
python-dotenv==1.0.1